contact us page in JS
